added news function
