+++ App.tsx
@@ -1,18 +1,10 @@
 import { NavigationContainer } from "@react-navigation/native";
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import home from "./component/home";
-
-export type RootStackParamList = {
-  Home: undefined;
-};
-const Stack = createNativeStackNavigator<RootStackParamList>();
+import AppNavigator from "./AppNavigator";
 
 export default function App() {
   return (
     <NavigationContainer>
-      <Stack.Navigator initialRouteName="Home">
-        <Stack.Screen name="Home" component={home} />
-      </Stack.Navigator>
+      <AppNavigator />
     </NavigationContainer>
   );
 }
