+++ package.json
@@ -12,6 +12,7 @@
     "@expo/metro-runtime": "~5.0.4",
     "@react-navigation/native": "^7.1.9",
     "@react-navigation/native-stack": "^7.3.13",
+    "@react-navigation/stack": "^7.3.2",
     "axios": "^1.9.0",
     "expo": "~53.0.9",
     "expo-constants": "~17.1.6",
@@ -23,7 +24,8 @@
     "react-native-dotenv": "^3.4.11",
     "react-native-safe-area-context": "5.4.0",
     "react-native-screens": "~4.10.0",
-    "react-native-web": "^0.20.0"
+    "react-native-web": "^0.20.0",
+    "react-native-webview": "13.13.5"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
